--- app/models/cfe/submission.rb
@@ -1,9 +1,0 @@
-module CFE
-  class Submission < ApplicationRecord
-    include CFESubmissionStateMachine
-
-    belongs_to :legal_aid_application
-    has_many :submission_histories, -> { order(created_at: :asc) }
-    has_one :result
-  end
-end
