--- app/models/cfe/submission_history.rb
@@ -1,5 +1,0 @@
-module CFE
-  class SubmissionHistory < ApplicationRecord
-    belongs_to :submission
-  end
-end
