+++ app/models/cfe/v1/submission.rb
@@ -1,0 +1,11 @@
+module CFE
+  module V1
+    class Submission < ApplicationRecord
+      include CFESubmissionStateMachine
+
+      belongs_to :legal_aid_application
+      has_many :submission_histories, -> { order(created_at: :asc) }
+      has_one :result
+    end
+  end
+end
