+++ app/models/cfe/v1/submission_history.rb
@@ -1,0 +1,7 @@
+module CFE
+  module V1
+    class SubmissionHistory < ApplicationRecord
+      belongs_to :submission
+    end
+  end
+end
