+++ app/models/cfe/v2/submission_history.rb
@@ -1,0 +1,7 @@
+module CFE
+  module V2
+    class SubmissionHistory < ApplicationRecord
+      belongs_to :submission
+    end
+  end
+end
