+++ spec/factories/cfe_submissions.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :cfe_submission, class: CFE::Submission do
+  factory :cfe_submission, class: CFE::V1::Submission do
     legal_aid_application { create :legal_aid_application }
     assessment_id { SecureRandom.uuid }
   end
