+++ spec/services/cfe/create_applicant_service_spec.rb
@@ -46,7 +46,7 @@ module CFE
             expect {
               CreateApplicantService.call(submission)
             }.to change { submission.submission_histories.count }.by 1
-            history = CFE::SubmissionHistory.last
+            history = CFE::V1::SubmissionHistory.last
             expect(history.submission_id).to eq submission.id
             expect(history.url).to eq service.cfe_url
             expect(history.http_method).to eq 'POST'
